Refactoring Code in VBA
PROJECT OVERVIEW
The purpose of this analysis is to refactor the module 2 solution code in order for it to run faster and shorter than it did. It must be made to loop through all of the stock data one time to collect information on each stock's total daily volume and the return. Moreover, the new analysis run time and evaluation must be compared to the run time of the subroutine before refactoring.
RESULTS
In 2017, TERP had a negative low percentage of return as indicated by "red". In 2018 however, only ENPR and RUN had positive percentage return while others had negative percentage. almost all of the stocks had a negative percent return. 

Summary
* ADVANTAGE
It reduces execution time hugely and somewhat more efficient in nature.

* DISADVANTAGE
Codes that are complex can be difficult to arrange. Also, a change to a line of code may alter the whole result.

PROS AND CONS
Codes can easily be changed, understood and maintained if attentions are paid to the refactoring processes.